requirements-vercel.txt
# Vercel specific requirements - optimized for serverless
Flask==3.0.0
flask-cors==4.0.0
tensorflow-cpu==2.15.0
pillow==10.0.0
numpy==1.24.3
requests==2.31.0
